pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "BLRI"
authors = [
  { name="Ross Donnachie", email="code@radonn.co.za" },
]
description = "Breakthrough Listen Radio Interferometry Python package (pronounced `blurry`)."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
    "Operating System :: OS Independent",
]
dependencies = [
  "h5py",
  "tomli",
  "pyerfa",
  "guppi @ git+https://github.com/MydonSolutions/guppi@v2.0",
  "pyproj",
  "pyyaml",
  "pydantic",
  "numpy"
]
dynamic = ["version"]

[tool.setuptools_scm]
version_file = "src/blri/_version.py"

[tool.setuptools]
zip-safe = false

[project.scripts]
blri_pycorr = "blri.entrypoints.correlate:correlate_cli"
blri_uvh5_diff = "blri.entrypoints.uvh5:diff"
blri_bfr5gen = "blri.entrypoints.bfr5:generate_for_raw"
blri_bfr5gen_raster = "blri.entrypoints.bfr5:generate_raster_for_raw"